fixed some annoying cast warnings

+++ Source/Base/Base/OSGBaseFunctions.inl
@@ -1346,8 +1346,8 @@ void osgsleep(UInt32 millisecs)
     struct timespec req;
     int ns;
     
-    req.tv_sec = millisecs / 1000;
-    req.tv_nsec = (millisecs % 1000) * 1000;
+    req.tv_sec = static_cast<long>(millisecs / 1000);
+    req.tv_nsec = static_cast<long>((millisecs % 1000) * 1000);
     
     while((req.tv_sec > 0 || req.tv_nsec > 0) &&
           (ns = nanosleep(&req, &req)) < 0)
